2/23/2016  I have changed the temperatures from double to int for simplicity.

+++ funnyWeekend/src/byui/CIT260/funnyWeekend/control/CalculationControl.java
@@ -11,12 +11,12 @@ package byui.CIT260.funnyWeekend.control;
  */
 public class CalculationControl {
     /**This calculates an average, a total, and converts a temperature in F t*/
-    public static double convertTemp(double tempF){
+    public static int convertTemp(int tempF){
         if (tempF<-50)
             return -100;
         if (tempF>150)
             return -200;
-        double tempC;
+        int tempC;
         tempC = ((tempF-32)*5)/9;
         
         return tempC;
